requirements.txt
numpy
opencv-python
mediapipe
torch
torchvision
pillow
rembg
streamlit